fix: "파사드 패턴 도입"
- Room, Reservation 파사드로 변경

Resolves: #147
