+++ src/main/java/com/example/airdns/domain/restschedule/repository/RestScheduleRepository.java
@@ -11,7 +11,7 @@ public interface RestScheduleRepository extends JpaRepository<RestSchedule, Long
     /**
      * 시작시간과 종료시간 사이에 있는 데이터가 있는지 확인
      */
-    Optional<RestSchedule> findFirstByEndTimeGreaterThanAndStartTimeLessThan(LocalDateTime startDate, LocalDateTime endDate);
+//    Optional<RestSchedule> findFirstByEndTimeGreaterThanAndStartTimeLessThan(LocalDateTime startDate, LocalDateTime endDate);
 
     Optional<RestSchedule> findFirstByRoomsIdAndStartTimeBeforeAndEndTimeAfter(Long roomsId, LocalDateTime checkOut, LocalDateTime checkIn);
 }
