+++ src/main/java/com/example/airdns/domain/restschedule/service/RestScheduleService.java
@@ -13,4 +13,5 @@ public interface RestScheduleService {
 
     void deleteRestSchedule(Long imagesId, Rooms rooms);
 
+    boolean hasRestScheduleInRoomBetweenTimes(Long roomsId, LocalDateTime startDate, LocalDateTime endDate);
 }
